refactor(tests): migrate inline test modules to external test files
- Move compaction, lfuk, hit_rate, lfu, lru eviction tests to dns_cache_test.rs
- Move chain.rs tests to adjacent chain_test.rs (private split_domain access)
- Move validator.rs tests to adjacent validator_test.rs (private verify_rrset_signatures)
- Move udp.rs tests to adjacent udp_test.rs (private validate_response_id)
- Use #[cfg(test)] #[path = "..."] mod tests; for co-located files needing private access
