+++ crates/infrastructure/src/dns/cache/eviction/lfuk.rs
@@ -45,70 +45,3 @@ impl EvictionPolicy for LfukPolicy {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::dns::cache::data::{CachedData, DnssecStatus};
-    use ferrous_dns_domain::RecordType;
-    use std::net::IpAddr;
-    use std::sync::Arc;
-
-    fn make_record_with_hits(hits: u64) -> crate::dns::cache::record::CachedRecord {
-        let record = crate::dns::cache::record::CachedRecord::new(
-            CachedData::IpAddresses(Arc::new(vec!["1.1.1.1".parse::<IpAddr>().unwrap()])),
-            300,
-            RecordType::A,
-            Some(DnssecStatus::Unknown),
-        );
-        for _ in 0..hits {
-            record.record_hit();
-        }
-        record
-    }
-
-    #[test]
-    fn test_lfuk_score_zero_hits_returns_bootstrap_score() {
-        let policy = LfukPolicy {
-            min_lfuk_score: 1.5,
-            k_value: 0.5,
-        };
-        let record = make_record_with_hits(0);
-        let score = policy.compute_score(&record, 1_000_000);
-        assert_eq!(
-            score, 1.5,
-            "Entries com hits=0 devem receber bootstrap score igual a min_lfuk_score"
-        );
-    }
-
-    #[test]
-    fn test_lfuk_score_below_min_is_negative_after_first_hit() {
-        let policy = LfukPolicy {
-            min_lfuk_score: 100.0,
-            k_value: 0.5,
-        };
-        // 1 hit numa entrada muito antiga → score calculado será << min_lfuk_score
-        let record = make_record_with_hits(1);
-        let score = policy.compute_score(&record, 1_000_000);
-        assert!(
-            score < 0.0,
-            "Entry com poucos hits e alto min_lfuk_score deve ter score negativo: {}",
-            score
-        );
-    }
-
-    #[test]
-    fn test_lfuk_score_with_many_hits_and_recent_access() {
-        let policy = LfukPolicy {
-            min_lfuk_score: 0.0,
-            k_value: 0.5,
-        };
-        let record = make_record_with_hits(20);
-        let now_secs = crate::dns::cache::coarse_clock::coarse_now_secs();
-        let score = policy.compute_score(&record, now_secs);
-        assert!(
-            score >= 0.0,
-            "Entrada com muitos hits recentes deve ter score >= 0"
-        );
-    }
-}
